chore(parser): move WithEnumExtension to openapi
