+++ src/openApi/common/interfaces/WithEnumExtension.ts
@@ -1,0 +1,4 @@
+export interface WithEnumExtension {
+    'x-enum-descriptions'?: ReadonlyArray<string>;
+    'x-enum-varnames'?: ReadonlyArray<string>;
+}
